+++ changelog.md
@@ -7,6 +7,7 @@ Because of that, some options may have reset to their default value
 
 - Added Bloodlust mode (TBC)
 - Added Grounding Totem mode
+- Added Battle Rez mode
 - Testing time increased to 60 minutes (up from 10 minutes)
 - Razuvious mode no longer exists, it was hidden since forever anyway