+++ changelog.md
@@ -1,6 +1,6 @@
 ## SilentRotate Changelog
 
-#### v0.8.0-beta (2021-xx-yy)
+#### v0.8.0-beta (2021-10-15)
 
 The code was reworked to improve performance and make it easier to maintain
 Because of that, some options may have reset to their default value
@@ -10,6 +10,7 @@ Because of that, some options may have reset to their default value
 - Added Battle Rez mode
 - Testing time increased to 60 minutes (up from 10 minutes)
 - Razuvious mode no longer exists, it was hidden since forever anyway
+- Support for Multi-TOC files (Vanilla and TBC)
 
 #### v0.7.1-beta (2021-09-19)